Fix intermittent bug where tokens were being incorrectly identified as Identifiers
The regex for matching an Identifier used to be /.*/, which meant that
if the parser looked at it first, any token would become an identifier.
Not sure what was happening with the loading order to make this occur,
but every once in a while (run the --failure-rate loop for fun times)
it would id incorrectly. I am now explicitly excluding
Assignment and Literal token matchers here.